requirements.txt
dlib>=19.4
numpy>=1.13
opencv-python>=3.2
tensorflow-gpu=>1.12
